refactor & executor skeleton
