+++ include/mori/io/common.hpp
