+++ include/mori/io/io.hpp
@@ -21,7 +21,7 @@
 // SOFTWARE.
 #pragma once
 
+#include "mori/io/common.hpp"
 #include "mori/io/engine.hpp"
 #include "mori/io/enum.hpp"
 #include "mori/io/logging.hpp"
-#include "mori/io/meta_data.hpp"
