+++ src/io/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_library(mori_io SHARED engine.cpp rdma/protocol.cpp
-                           rdma/backend_impl_v1.cpp rdma/executor.cpp)
+add_library(mori_io SHARED engine.cpp rdma/protocol.cpp rdma/backend_impl.cpp
+                           rdma/executor.cpp rdma/common.cpp)
 
 target_include_directories(mori_io PUBLIC ${CMAKE_SOURCE_DIR}/include)
 target_include_directories(mori_io PUBLIC ${CMAKE_SOURCE_DIR})
