+++ src/io/rdma/protocol.hpp
@@ -23,9 +23,9 @@
 
 #include <msgpack.hpp>
 
-#include "mori/io/meta_data.hpp"
+#include "mori/io/common.hpp"
 #include "mori/io/msgpack_adaptor.hpp"
-#include "src/io/rdma/backend_impl_v1.hpp"
+#include "src/io/rdma/backend_impl.hpp"
 
 namespace mori {
 namespace io {
